Note that this implements the "Vendor Branches" strategy described in the Subversion book at
http://svnbook.red-bean.com/nightly/en/svn.branchmerge.basicmerging.html
http://svnbook.red-bean.com/nightly/en/svn.advanced.vendorbr.html

Closes #70.



git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@324 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ ExtLibs/wxWidgets/src/common/fileconf.cpp
@@ -146,7 +146,7 @@ private:
 
   wxString      m_strName,      // entry name
                 m_strValue;     //       value
-  bool          m_bImmutable:1, // can be overriden locally?
+  bool          m_bImmutable:1, // can be overridden locally?
                 m_bHasValue:1;  // set after first call to SetValue()
 
   int           m_nLine;        // used if m_pLine == NULL only
@@ -661,7 +661,7 @@ void wxFileConfig::Parse(const wxTextBuffer& buffer, bool bLocal)
           // which is exactly what we want.
           else if ( !bLocal || pEntry->IsLocal() ) {
             wxLogWarning(_("file '%s', line %d: key '%s' was first found at line %d."),
-                         buffer.GetName(), n + 1, strKey.c_str(), pEntry->Line());
+                         buffer.GetName(), (int)n + 1, strKey.c_str(), pEntry->Line());
 
           }
         }
